# internal/api/network/gen/network.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package gen

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
	openapi_types "github.com/oapi-codegen/runtime/types"
	externalRef0 "github.com/retr0h/osapi/internal/api/common/gen"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// Defines values for DNSUpdateResultItemStatus.
const (
	Failed DNSUpdateResultItemStatus = "failed"
	Ok     DNSUpdateResultItemStatus = "ok"
)

// DNSConfigCollectionResponse defines model for DNSConfigCollectionResponse.
type DNSConfigCollectionResponse struct {
	// JobId The job ID used to process this request.
	JobId   *openapi_types.UUID `json:"job_id,omitempty"`
	Results []DNSConfigResponse `json:"results"`
}

// DNSConfigResponse defines model for DNSConfigResponse.
type DNSConfigResponse struct {
	// Hostname The hostname of the worker that served this DNS config.
	Hostname string `json:"hostname"`

	// SearchDomains List of search domains.
	SearchDomains *[]string `json:"search_domains,omitempty"`

	// Servers List of configured DNS servers.
	Servers *[]string `json:"servers,omitempty"`
}

// DNSConfigUpdateRequest defines model for DNSConfigUpdateRequest.
type DNSConfigUpdateRequest struct {
	// InterfaceName The name of the network interface to apply DNS configuration to. Must only contain letters and numbers.
	InterfaceName string `json:"interface_name" validate:"required,alphanum"`

	// SearchDomains New list of search domains to configure.
	SearchDomains *[]string `json:"search_domains,omitempty" validate:"required_without=Servers,omitempty,dive,hostname,min=1"`

	// Servers New list of DNS servers to configure.
	Servers *[]string `json:"servers,omitempty" validate:"required_without=SearchDomains,omitempty,dive,ip,min=1"`
}

// DNSUpdateCollectionResponse defines model for DNSUpdateCollectionResponse.
type DNSUpdateCollectionResponse struct {
	// JobId The job ID used to process this request.
	JobId   *openapi_types.UUID   `json:"job_id,omitempty"`
	Results []DNSUpdateResultItem `json:"results"`
}

// DNSUpdateResultItem defines model for DNSUpdateResultItem.
type DNSUpdateResultItem struct {
	Error    *string                   `json:"error,omitempty"`
	Hostname string                    `json:"hostname"`
	Status   DNSUpdateResultItemStatus `json:"status"`
}

// DNSUpdateResultItemStatus defines model for DNSUpdateResultItem.Status.
type DNSUpdateResultItemStatus string

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse = externalRef0.ErrorResponse

// PingCollectionResponse defines model for PingCollectionResponse.
type PingCollectionResponse struct {
	// JobId The job ID used to process this request.
	JobId   *openapi_types.UUID `json:"job_id,omitempty"`
	Results []PingResponse      `json:"results"`
}

// PingResponse defines model for PingResponse.
type PingResponse struct {
	// AvgRtt Average round-trip time as a string in Go's time.Duration format.
	AvgRtt *string `json:"avg_rtt,omitempty"`

	// Hostname The hostname of the worker that executed the ping.
	Hostname string `json:"hostname"`

	// MaxRtt Maximum round-trip time as a string in Go's time.Duration format.
	MaxRtt *string `json:"max_rtt,omitempty"`

	// MinRtt Minimum round-trip time as a string in Go's time.Duration format.
	MinRtt *string `json:"min_rtt,omitempty"`

	// PacketLoss Percentage of packet loss.
	PacketLoss *float64 `json:"packet_loss,omitempty"`

	// PacketsReceived Number of packets received.
	PacketsReceived *int `json:"packets_received,omitempty"`

	// PacketsSent Number of packets sent.
	PacketsSent *int `json:"packets_sent,omitempty"`
}

// PutNetworkDNSParams defines parameters for PutNetworkDNS.
type PutNetworkDNSParams struct {
	// TargetHostname Target: _any (load-balanced), _all (broadcast), hostname (direct), or key:value (label group, e.g., group:web.dev).
	TargetHostname *string `form:"target_hostname,omitempty" json:"target_hostname,omitempty"`
}

// GetNetworkDNSByInterfaceParams defines parameters for GetNetworkDNSByInterface.
type GetNetworkDNSByInterfaceParams struct {
	// TargetHostname Target: _any (load-balanced), _all (broadcast), hostname (direct), or key:value (label group, e.g., group:web.dev).
	TargetHostname *string `form:"target_hostname,omitempty" json:"target_hostname,omitempty"`
}

// PostNetworkPingJSONBody defines parameters for PostNetworkPing.
type PostNetworkPingJSONBody struct {
	// Address The IP address of the server to ping. Supports both IPv4 and IPv6.
	Address string `json:"address" validate:"required,ip"`
}

// PostNetworkPingParams defines parameters for PostNetworkPing.
type PostNetworkPingParams struct {
	// TargetHostname Target: _any (load-balanced), _all (broadcast), hostname (direct), or key:value (label group, e.g., group:web.dev).
	TargetHostname *string `form:"target_hostname,omitempty" json:"target_hostname,omitempty"`
}

// PutNetworkDNSJSONRequestBody defines body for PutNetworkDNS for application/json ContentType.
type PutNetworkDNSJSONRequestBody = DNSConfigUpdateRequest

// PostNetworkPingJSONRequestBody defines body for PostNetworkPing for application/json ContentType.
type PostNetworkPingJSONRequestBody PostNetworkPingJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Update DNS servers
	// (PUT /network/dns)
	PutNetworkDNS(ctx echo.Context, params PutNetworkDNSParams) error
	// List DNS servers
	// (GET /network/dns/{interfaceName})
	GetNetworkDNSByInterface(ctx echo.Context, interfaceName string, params GetNetworkDNSByInterfaceParams) error
	// Ping a remote server
	// (POST /network/ping)
	PostNetworkPing(ctx echo.Context, params PostNetworkPingParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PutNetworkDNS converts echo context to params.
func (w *ServerInterfaceWrapper) PutNetworkDNS(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{"write"})

	// Parameter object where we will unmarshal all parameters from the context
	var params PutNetworkDNSParams
	// ------------- Optional query parameter "target_hostname" -------------

	err = runtime.BindQueryParameter("form", true, false, "target_hostname", ctx.QueryParams(), &params.TargetHostname)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter target_hostname: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PutNetworkDNS(ctx, params)
	return err
}

// GetNetworkDNSByInterface converts echo context to params.
func (w *ServerInterfaceWrapper) GetNetworkDNSByInterface(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "interfaceName" -------------
	var interfaceName string

	err = runtime.BindStyledParameterWithOptions("simple", "interfaceName", ctx.Param("interfaceName"), &interfaceName, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter interfaceName: %s", err))
	}

	ctx.Set(BearerAuthScopes, []string{"read"})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetNetworkDNSByInterfaceParams
	// ------------- Optional query parameter "target_hostname" -------------

	err = runtime.BindQueryParameter("form", true, false, "target_hostname", ctx.QueryParams(), &params.TargetHostname)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter target_hostname: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetNetworkDNSByInterface(ctx, interfaceName, params)
	return err
}

// PostNetworkPing converts echo context to params.
func (w *ServerInterfaceWrapper) PostNetworkPing(ctx echo.Context) error {
	var err error

	ctx.Set(BearerAuthScopes, []string{"write"})

	// Parameter object where we will unmarshal all parameters from the context
	var params PostNetworkPingParams
	// ------------- Optional query parameter "target_hostname" -------------

	err = runtime.BindQueryParameter("form", true, false, "target_hostname", ctx.QueryParams(), &params.TargetHostname)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter target_hostname: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostNetworkPing(ctx, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.PUT(baseURL+"/network/dns", wrapper.PutNetworkDNS)
	router.GET(baseURL+"/network/dns/:interfaceName", wrapper.GetNetworkDNSByInterface)
	router.POST(baseURL+"/network/ping", wrapper.PostNetworkPing)

}

type PutNetworkDNSRequestObject struct {
	Params PutNetworkDNSParams
	Body   *PutNetworkDNSJSONRequestBody
}

type PutNetworkDNSResponseObject interface {
	VisitPutNetworkDNSResponse(w http.ResponseWriter) error
}

type PutNetworkDNS202JSONResponse DNSUpdateCollectionResponse

func (response PutNetworkDNS202JSONResponse) VisitPutNetworkDNSResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(202)

	return json.NewEncoder(w).Encode(response)
}

type PutNetworkDNS400JSONResponse externalRef0.ErrorResponse

func (response PutNetworkDNS400JSONResponse) VisitPutNetworkDNSResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PutNetworkDNS401JSONResponse externalRef0.ErrorResponse

func (response PutNetworkDNS401JSONResponse) VisitPutNetworkDNSResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type PutNetworkDNS403JSONResponse externalRef0.ErrorResponse

func (response PutNetworkDNS403JSONResponse) VisitPutNetworkDNSResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)

	return json.NewEncoder(w).Encode(response)
}

type PutNetworkDNS500JSONResponse externalRef0.ErrorResponse

func (response PutNetworkDNS500JSONResponse) VisitPutNetworkDNSResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetNetworkDNSByInterfaceRequestObject struct {
	InterfaceName string `json:"interfaceName"`
	Params        GetNetworkDNSByInterfaceParams
}

type GetNetworkDNSByInterfaceResponseObject interface {
	VisitGetNetworkDNSByInterfaceResponse(w http.ResponseWriter) error
}

type GetNetworkDNSByInterface200JSONResponse DNSConfigCollectionResponse

func (response GetNetworkDNSByInterface200JSONResponse) VisitGetNetworkDNSByInterfaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetNetworkDNSByInterface400JSONResponse externalRef0.ErrorResponse

func (response GetNetworkDNSByInterface400JSONResponse) VisitGetNetworkDNSByInterfaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetNetworkDNSByInterface401JSONResponse externalRef0.ErrorResponse

func (response GetNetworkDNSByInterface401JSONResponse) VisitGetNetworkDNSByInterfaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type GetNetworkDNSByInterface403JSONResponse externalRef0.ErrorResponse

func (response GetNetworkDNSByInterface403JSONResponse) VisitGetNetworkDNSByInterfaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)

	return json.NewEncoder(w).Encode(response)
}

type GetNetworkDNSByInterface500JSONResponse externalRef0.ErrorResponse

func (response GetNetworkDNSByInterface500JSONResponse) VisitGetNetworkDNSByInterfaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type PostNetworkPingRequestObject struct {
	Params PostNetworkPingParams
	Body   *PostNetworkPingJSONRequestBody
}

type PostNetworkPingResponseObject interface {
	VisitPostNetworkPingResponse(w http.ResponseWriter) error
}

type PostNetworkPing200JSONResponse PingCollectionResponse

func (response PostNetworkPing200JSONResponse) VisitPostNetworkPingResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type PostNetworkPing400JSONResponse externalRef0.ErrorResponse

func (response PostNetworkPing400JSONResponse) VisitPostNetworkPingResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PostNetworkPing401JSONResponse externalRef0.ErrorResponse

func (response PostNetworkPing401JSONResponse) VisitPostNetworkPingResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)

	return json.NewEncoder(w).Encode(response)
}

type PostNetworkPing403JSONResponse externalRef0.ErrorResponse

func (response PostNetworkPing403JSONResponse) VisitPostNetworkPingResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(403)

	return json.NewEncoder(w).Encode(response)
}

type PostNetworkPing500JSONResponse externalRef0.ErrorResponse

func (response PostNetworkPing500JSONResponse) VisitPostNetworkPingResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Update DNS servers
	// (PUT /network/dns)
	PutNetworkDNS(ctx context.Context, request PutNetworkDNSRequestObject) (PutNetworkDNSResponseObject, error)
	// List DNS servers
	// (GET /network/dns/{interfaceName})
	GetNetworkDNSByInterface(ctx context.Context, request GetNetworkDNSByInterfaceRequestObject) (GetNetworkDNSByInterfaceResponseObject, error)
	// Ping a remote server
	// (POST /network/ping)
	PostNetworkPing(ctx context.Context, request PostNetworkPingRequestObject) (PostNetworkPingResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// PutNetworkDNS operation middleware
func (sh *strictHandler) PutNetworkDNS(ctx echo.Context, params PutNetworkDNSParams) error {
	var request PutNetworkDNSRequestObject

	request.Params = params

	var body PutNetworkDNSJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PutNetworkDNS(ctx.Request().Context(), request.(PutNetworkDNSRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PutNetworkDNS")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PutNetworkDNSResponseObject); ok {
		return validResponse.VisitPutNetworkDNSResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetNetworkDNSByInterface operation middleware
func (sh *strictHandler) GetNetworkDNSByInterface(ctx echo.Context, interfaceName string, params GetNetworkDNSByInterfaceParams) error {
	var request GetNetworkDNSByInterfaceRequestObject

	request.InterfaceName = interfaceName
	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetNetworkDNSByInterface(ctx.Request().Context(), request.(GetNetworkDNSByInterfaceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetNetworkDNSByInterface")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetNetworkDNSByInterfaceResponseObject); ok {
		return validResponse.VisitGetNetworkDNSByInterfaceResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PostNetworkPing operation middleware
func (sh *strictHandler) PostNetworkPing(ctx echo.Context, params PostNetworkPingParams) error {
	var request PostNetworkPingRequestObject

	request.Params = params

	var body PostNetworkPingJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostNetworkPing(ctx.Request().Context(), request.(PostNetworkPingRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostNetworkPing")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostNetworkPingResponseObject); ok {
		return validResponse.VisitPostNetworkPingResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}