# internal/client/gen/api.yaml
openapi: 3.0.0
info:
  title: OSAPI - A CRUD API for managing Linux systems
  description: |
    OSAPI is a lightweight, minimalistic RESTful API designed to manage and
    configure Linux systems as appliances. The API provides comprehensive
    management capabilities, including reading the current state of the system
    and making privileged, asynchronous changes to match a desired configuration
    state.

    OSAPI is built with pluggability and extensibility in mind, allowing it to
    adapt to different Linux distributions and environments by implementing
    custom handlers for various system operations.

    Key Features:
    - **RESTful API Design:** A RESTful approach for managing Linux systems,
      enabling clear and organized endpoints for system configurations,
      diagnostics, and state management.
    - **Pluggable Architecture:** Easily extend or replace components to suit
      different Linux environments. This flexibility allows users to customize
      behavior based on specific distribution requirements or operational needs.
    - **Job System for Privileged Actions:** Asynchronous job execution via
      NATS JetStream to perform system modifications securely, ensuring that
      changes are managed in a controlled and orderly manner.
    - **System Information Gathering:** Provides detailed system insights, such
      as OS details, hostname, and other environment-specific information,
      facilitating diagnostics and monitoring.
...
  contact:
...
paths:
...
  /network/dns/{interfaceName}:
    servers: []
    get:
      summary: List DNS servers
      description: >-
        Retrieve the list of currently configured DNS servers for a specific
        network interface.
      tags:
        - Network_Management_API_dns_operations
      operationId: GetNetworkDNSByInterface
      security:
        - BearerAuth:
            - read
      parameters:
        - name: interfaceName
          in: path
          required: true
          schema:
            type: string
            x-oapi-codegen-extra-tags:
              validate: required,alphanum
          description: >
            The name of the network interface to retrieve DNS configuration for.
            Must only contain letters and numbers.
        - name: target_hostname
          in: query
          required: false
          schema:
            type: string
            default: _any
          description: >
            Target: _any (load-balanced), _all (broadcast), hostname (direct),
            or key:value (label group, e.g., group:web.dev).
      responses:
        '200':
          description: List of DNS servers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DNSConfigCollectionResponse'
        '400':
          description: Invalid interface name provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error retrieving DNS servers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /network/dns:
    servers: []
    put:
      summary: Update DNS servers
      description: Update the system's DNS server configuration.
      tags:
        - Network_Management_API_dns_operations
      operationId: PutNetworkDNS
      security:
        - BearerAuth:
            - write
      parameters:
        - name: target_hostname
          in: query
          required: false
          schema:
            type: string
            default: _any
          description: >
            Target: _any (load-balanced), _all (broadcast), hostname (direct),
            or key:value (label group, e.g., group:web.dev).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DNSConfigUpdateRequest'
      responses:
        '202':
          description: DNS servers update successfully accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DNSUpdateCollectionResponse'
        '400':
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - API key required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error updating DNS servers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A description of the error that occurred.
          example: Failed to retrieve status.
        details:
          type: string
          description: Additional details about the error.
          example: Failed due to network timeout.
        code:
          type: integer
          description: The error code.
          example: 500
...
    DNSConfigResponse:
      type: object
      properties:
        hostname:
          type: string
          description: The hostname of the worker that served this DNS config.
        servers:
          type: array
          description: List of configured DNS servers.
          items:
            type: string
            description: IPv4 or IPv6 address of the DNS server.
        search_domains:
          type: array
          description: List of search domains.
          items:
            type: string
      required:
        - hostname
    DNSConfigCollectionResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
          description: The job ID used to process this request.
          example: 550e8400-e29b-41d4-a716-446655440000
        results:
          type: array
          items:
            $ref: '#/components/schemas/DNSConfigResponse'
      required:
        - results
    DNSUpdateResultItem:
      type: object
      properties:
        hostname:
          type: string
        status:
          type: string
          enum:
            - ok
            - failed
        error:
          type: string
      required:
        - hostname
        - status
    DNSUpdateCollectionResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
          description: The job ID used to process this request.
          example: 550e8400-e29b-41d4-a716-446655440000
        results:
          type: array
          items:
            $ref: '#/components/schemas/DNSUpdateResultItem'
      required:
        - results
    DNSConfigUpdateRequest:
      type: object
      properties:
        servers:
          type: array
          x-oapi-codegen-extra-tags:
            validate: required_without=SearchDomains,omitempty,dive,ip,min=1
          description: New list of DNS servers to configure.
          items:
            type: string
            description: IPv4 or IPv6 address of the DNS server.
        search_domains:
          type: array
          x-oapi-codegen-extra-tags:
            validate: required_without=Servers,omitempty,dive,hostname,min=1
          description: New list of search domains to configure.
          items:
            type: string
        interface_name:
          type: string
          x-oapi-codegen-extra-tags:
            validate: required,alphanum
          description: >
            The name of the network interface to apply DNS configuration to.
            Must only contain letters and numbers.
      required:
        - interface_name
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...